Editor GUI Fixes (#15)
* Repaint editor GUIs on shortcut

* Editor variable changes saved properly

Before changing a variable would not be detected and so not saved/serialized in some cases.

* Use callback attributes to repaint editors

This would not work with the `upm-support` branch #14 as the attributes are not searched for in the attribute that these scripts are in by default.

* Allow multiple un/loadTrigger attributes, use action to trigger repaints

* Remove SceneManagement, it's not required

The scene is automatically set as dirty when the gui target is

* Fixes to gui with enableInEdit mode

Admittedly this should have been in #12. However, here are the fixes anyways.

GUI buttons to un/load show in edit mode if enableInEditMode is true. Disabling and re-enabling the DllManipulatorScript works now.

* Properly reset custom triggers (fix duplicates)

Previously triggers would not be cleared properly and so be duplicated. (Static variables seem to persist between entering/exiting playmode)

* Custom triggers optional execute on main thread

Allow custom triggers to be executed on the main thread in a queue. We need this internally to repaint the editor GUI as it uses the Unity API.

* Update README.md

* Mark scene dirty only if options changed

* Bug fix, resetting when not initialized

Initialize() is not always called so we should not always reset.

* small fix

* Review fixes

Also RegisterTriggerMethod receives attribute instead of bool, 
Fix naming in DllManipulatorOptions

* Add comments