Add upm support (#14)
* Add upm support

Allow users to add this to the Packages folder for it to be used as an embedded package
Users can then nicely add the repo as a submodule

* Add meta files

* Potential fix for assemblyPaths using their manifest name

* Fallback to windows #if

* Rename asmdefs

* Improve finding assemblies, always search in our assembly

Renamed Editor asmdef

* assemblyPaths small fixes, editor gui adapted

EditorGUI for the target assemblies is now just a string array. Removed code to find _allKnownAssemblies.
assemblyPaths renamed to assemblyNames is now a list.
By default only Assembly-CSharp and the Editor variant is targeted. The assembly containing this tool is always included so we find the attributes.

* Improve UI, dropdowns, add _allKnownAssemblies back

Simplified finding all assemblies, moved to RefreshAllKnownAssemblies()
Unity Assemblies are ignored by default.
Option to add all known assemblies.
Increased refresh time to 5 seconds from 1.

* Add gui headers, info box

* Update README.md

* Partial revert "assemblyPaths small fixes, editor gui adapted", remove UnityEditorInternal dependency

Use original code to find all assemblies, use `Path.GetFileNameWithoutExtension` instead of `PathUtils.NormallizeUnityAssemblyPath`.

* Use string[] for Options.assemblyNames

* Changed package.json

* Bug fix, disable whole gui in play mode

* Revert "Use string[] for Options.assemblyNames"

This reverts commit 8e1d0e370bdbeb533879ec602e39a0995ff73283.

* Bug fix, hide errors for not finding default assemblies

Some project may not have any editor scripts so will always get an error as Assembly-CSharp-Editor does not exist.

* Merge fixes

* Review changes

* Removed upm author field