Edit options when initialized, Fix cleanup when recompiling (#20)
* Allow editing options when initialized, DllManip has a separate Options copy to DllManipScript

When DllManipulator is initialized we pass a copy of the current DllManipulatorScript options. This allows us to keep editing the options when the manipulator is initialized/dlls are loaded as we are editing a copy. This is particularly useful when using enableInEditMode.

Changes are always applied at OnEnable as before. To allow modifying without play/stop a DllManipulatorScript.Reinitialize() has been added which can be called via GUI (see DetectOptionChanges) which unloads dlls and initializes with new options. GUI is only shown when there are changes.

* Fix OnDestroy not being called when recompiling

When recompiling only `OnDisable` is called, not `OnDestroy`. Now gets the callback for when the assembly is being reloaded, just when finished with compiling. This is called before `OnDisable`, so I just set a flag `_isRecompiling`.

Previously the native function `UnityPluginUnload` was not being called causing various bugs.

* Update scripts/Editor/DllManipulatorEditor.cs

Co-authored-by: mcpiroman <38111589+mcpiroman@users.noreply.github.com>

* Review changes

Also removed check if singletonInstance in Reinitialize as this should never occur

* Small fix

* DllManip.Options private set

Co-authored-by: mcpiroman <38111589+mcpiroman@users.noreply.github.com>